backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via ortools
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via aiohttp
attrs==25.1.0
    # via aiohttp
azure-common==1.1.28
    # via azure-maps-route
azure-core==1.32.0
    # via
    #   azure-mgmt-core
    #   msrest
azure-maps-route==1.0.0b3
    # via -r requirements.in
azure-mgmt-core==1.5.0
    # via azure-maps-route
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi[standard]==0.115.8
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
immutabledict==4.2.1
    # via ortools
isodate==0.7.2
    # via msrest
jinja2==3.1.5
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msrest==0.7.1
    # via azure-maps-route
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.3
    # via
    #   ortools
    #   pandas
oauthlib==3.2.2
    # via requests-oauthlib
ortools==9.11.4210
    # via -r requirements.in
pandas==2.2.3
    # via ortools
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.26.1
    # via ortools
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   -r requirements.in
    #   azure-core
    #   msrest
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via msrest
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
typer==0.15.1
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   fastapi
    #   multidict
    #   pydantic
    #   pydantic-core
    #   rich
    #   rich-toolkit
    #   typer
    #   uvicorn
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==14.2
    # via uvicorn
yarl==1.18.3
    # via aiohttp
